+++ controller/login/get-single-controller.go
@@ -3,6 +3,8 @@ package login
 import (
 	"encoding/base64"
 	"log"
+	"net/http"
+	"strconv"
 
 	"github.com/yakuter/gpass/controller/helper"
 	"github.com/yakuter/gpass/model"
@@ -18,11 +20,20 @@ func GetLogin(c *gin.Context) {
 	id := c.Params.ByName("id")
 	var login model.Login
 
-	if err := db.Where("id = ? ", id).First(&login).Error; err != nil {
-		log.Println(err)
-		c.AbortWithStatus(404)
+	idInt, err := strconv.Atoi(id)
+	if err != nil {
+		result := model.Result{"Error", err.Error()}
+		c.JSON(http.StatusBadRequest, result)
 		return
 	}
+
+	if err := db.Where("id = ? ", idInt).First(&login).Error; err != nil {
+		log.Println(err)
+		result := model.Result{"Error", err.Error()}
+		c.JSON(http.StatusNotFound, result)
+		return
+	}
+
 	passByte, _ := base64.StdEncoding.DecodeString(login.Password)
 	passB64 := helper.Decrypt(string(passByte[:]), config.Server.Passphrase)
 	login.Password = passB64